requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.2.16
sqlparse==0.5.1
typing-extensions==4.12.2
djangorestframework==3.15.2;python_version>"3"
docutils==0.21.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
PyJWT==2.10.1